creared BaseTest, extrzct locators in variables
